# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sankhya Support System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardSummary {
  openTickets: number;
  pendingReleases: number;
  activeClients: number;
  /** @nullable */
  lastSync: string | null;
  totalTickets: number;
  closedThisMonth: number;
}

export interface MonthlyCount {
  month: string;
  opened: number;
  closed: number;
}

export interface StatusBreakdown {
  label: string;
  count: number;
}

export interface ActivityLog {
  id: number;
  action: string;
  entity: string;
  description: string;
  /** @nullable */
  user?: string | null;
  createdAt: string;
}

export interface Ticket {
  id: number;
  title: string;
  description: string;
  priority: string;
  status: string;
  sector: string;
  openedBy: string;
  /** @nullable */
  assignedTo?: string | null;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  clientName?: string | null;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  closedAt?: string | null;
}

export interface TicketInput {
  title: string;
  description: string;
  priority: string;
  sector: string;
  openedBy: string;
  /** @nullable */
  assignedTo?: string | null;
  /** @nullable */
  clientId?: number | null;
}

export interface TicketUpdate {
  title?: string;
  description?: string;
  priority?: string;
  status?: string;
  sector?: string;
  /** @nullable */
  assignedTo?: string | null;
  /** @nullable */
  clientId?: number | null;
}

export interface Client {
  id: number;
  name: string;
  sankhyaCode: string;
  cnpj: string;
  phone: string;
  email: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface ClientInput {
  name: string;
  sankhyaCode: string;
  cnpj: string;
  phone: string;
  email: string;
  /** @nullable */
  notes?: string | null;
}

export interface Release {
  id: number;
  clientId: number;
  clientName: string;
  requestedBy: string;
  description: string;
  status: string;
  createdAt: string;
}

export interface ReleaseInput {
  clientId: number;
  requestedBy: string;
  description: string;
  status?: string;
}

export interface ReleaseUpdate {
  status?: string;
  description?: string;
}

export type SyncStatusTablesItem = {
  name: string;
  /** @nullable */
  lastSync: string | null;
  recordCount: number;
  status: string;
};

export interface SyncStatus {
  /** @nullable */
  lastSync?: string | null;
  tables: SyncStatusTablesItem[];
}

export type ListTicketsParams = {
  status?: string;
  priority?: string;
};