# lib/api-zod/src/generated/types/activityLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sankhya Support System API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityLog {
  id: number;
  action: string;
  entity: string;
  description: string;
  /** @nullable */
  user?: string | null;
  createdAt: Date;
}