# lib/api-zod/src/generated/types/clientInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sankhya Support System API
 * OpenAPI spec version: 0.1.0
 */

export interface ClientInput {
  name: string;
  sankhyaCode: string;
  cnpj: string;
  phone: string;
  email: string;
  /** @nullable */
  notes?: string | null;
}