# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sankhya Support System API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  openTickets: number;
  pendingReleases: number;
  activeClients: number;
  /** @nullable */
  lastSync: Date | null;
  totalTickets: number;
  closedThisMonth: number;
}