# lib/api-zod/src/generated/types/releaseInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sankhya Support System API
 * OpenAPI spec version: 0.1.0
 */

export interface ReleaseInput {
  clientId: number;
  requestedBy: string;
  description: string;
  status?: string;
}