# lib/api-zod/src/generated/types/releaseUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sankhya Support System API
 * OpenAPI spec version: 0.1.0
 */

export interface ReleaseUpdate {
  status?: string;
  description?: string;
}