# lib/api-zod/src/generated/types/ticket.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sankhya Support System API
 * OpenAPI spec version: 0.1.0
 */

export interface Ticket {
  id: number;
  title: string;
  description: string;
  priority: string;
  status: string;
  sector: string;
  openedBy: string;
  /** @nullable */
  assignedTo?: string | null;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  clientName?: string | null;
  createdAt: Date;
  updatedAt: Date;
  /** @nullable */
  closedAt?: Date | null;
}