# lib/api-zod/src/generated/types/ticketInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sankhya Support System API
 * OpenAPI spec version: 0.1.0
 */

export interface TicketInput {
  title: string;
  description: string;
  priority: string;
  sector: string;
  openedBy: string;
  /** @nullable */
  assignedTo?: string | null;
  /** @nullable */
  clientId?: number | null;
}