# lib/api-zod/src/generated/types/ticketUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sankhya Support System API
 * OpenAPI spec version: 0.1.0
 */

export interface TicketUpdate {
  title?: string;
  description?: string;
  priority?: string;
  status?: string;
  sector?: string;
  /** @nullable */
  assignedTo?: string | null;
  /** @nullable */
  clientId?: number | null;
}